backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
drf-yasg==1.21.7
gunicorn==21.2.0
whitenoise==6.6.0
